requirements.txt
appdirs==1.4.4
asgiref==3.2.7
Babel==2.8.0
Brotli==1.0.7
certifi==2020.4.5.1
distlib==0.3.0

django-brotli==0.2.0
django-ckeditor==6.0.0
django-countries==6.1.2
django-js-asset==1.2.2

django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
django-widget-tweaks==1.4.8
filelock==3.0.12
html5lib==1.1
phonenumbers==8.12.4
Pillow==7.1.2
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.44
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
django-crispy-forms==1.12.0
